update deps, update po template

+++ rnote-engine/src/strokes/vectorimage.rs
@@ -168,7 +168,8 @@ impl VectorImage {
         let mut svg_tree = usvg::Tree::from_str(svg_data, &usvg::Options::default())?;
         svg_tree.convert_text(&render::USVG_FONTDB);
         let svg_data = svg_tree.to_string(&xml_options);
-        let intrinsic_size = na::vector![svg_tree.size.width(), svg_tree.size.height()];
+        let intrinsic_size =
+            na::vector![svg_tree.size.width() as f64, svg_tree.size.height() as f64];
 
         let rectangle = if let Some(size) = size {
             Rectangle {
